requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
google-generativeai
python-dotenv
python-multipart
google-genai